Refactor: Refine generate methods
