+++ src/asm/visitor/mod.rs
@@ -1,1 +1,2 @@
 pub mod signature;
+pub mod class;
